test/regression/reftest/vendor/blink/fast/dom/Element/id-in-map.html

<html>
<head>
<script src="../../../resources/wpt-test-pre.js"></script>
<script src="../../../resources/js-test-pre.js"></script>
</head>
<body>

<map title="map" id="firstmap">

<script>
description("Test using id in map element.");
    
shouldBeNonNull('document.getElementById("firstmap")');
</script>
<script src="../../../resources/js-test-post.js"></script>
<script src="../../../resources/wpt-test-post.js"></script>
</body>
</html>
