third_party/blink/web_tests/fast/dom/Element/id-in-map-expected.txt

Test using id in map element.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS document.getElementById("firstmap") is non-null.
PASS successfullyParsed is true

TEST COMPLETE
